yt_chat: rename llm to core
